butterflyfx-core/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
ed25519==1.5
httpx==0.26.0
PyYAML==6.0.1
python-dotenv==1.0.1
